chore: separate handlers to separate file for native-packages
