+++ package/native-package/src/packages/triggerHaptic.ts
@@ -1,0 +1,8 @@
+import ReactNativeHapticFeedback from 'react-native-haptic-feedback';
+
+export const triggerHaptic = (method: ReactNativeHapticFeedback.HapticFeedbackTypes) => {
+  ReactNativeHapticFeedback.trigger(method, {
+    enableVibrateFallback: false,
+    ignoreAndroidSystemSettings: false,
+  });
+};
